Api key

+++ static/assets/waiter_v2.js
@@ -14,6 +14,7 @@ var didScroll;
 var lastScrollTop = 0;
 var delta = 10;
 var binge_mode;
+var is_mcp;
 var has_next_link;
 var should_redirect = true;
 var username;
@@ -53,6 +54,10 @@ function prepareDataTable($){
 }
 
 function storeVideoPosition(filename, video){
+    if(is_mcp){
+        return;
+    }
+
     var offset = Math.max(video.currentTime, 0);
     console.log("Attempting to store video position");
     jQuery.ajax({url: offsetUrl + guid + '/' + filename + '/',
@@ -65,6 +70,10 @@ function storeVideoPosition(filename, video){
 }
 
 function getVideoPosition(filename, guid, video){
+    if(is_mcp){
+        return;
+    }
+
     console.log("Attempting to get video position");
     if(localStorage.getItem(filename)){
         video.currentTime(localStorage.getItem(filename));
@@ -88,6 +97,10 @@ function getVideoPosition(filename, guid, video){
 }
 
 function clearVideoPosition(filename){
+    if(is_mcp){
+        return;
+    }
+
     console.log("Attempting to clear video position");
     jQuery.ajax({url: offsetUrl + guid + '/' + filename + '/',
                  type: 'DELETE',
@@ -101,7 +114,42 @@ function clearVideoPosition(filename){
             });
 }
 
+function _display_binge_redirect(){
+    var finishedElement = document.getElementById('viewedText');
+    var toggleElement = document.getElementById('bingewatch-btn');
+    var rowElement = document.getElementById('binge-watch-row');
+
+    if(binge_mode && has_next_link && should_redirect){
+        var pretext;
+
+        if(!is_mcp){
+            pretext = "Marking file viewed... Binge mode active! Preparing next video";
+        } else {
+            pretext = "Binge mode active! Preparing next video";
+        };
+
+        finishedElement.innerHTML = pretext + " <button class='btn btn-info play-controls' name='cancel-btn' id='cancel-btn' onclick='cancelBingeWatch();'><i class='bi-x-octagon-fill'></i> Cancel</button>";
+        if(toggleElement){
+            toggleElement.onclick = function(){};
+            toggleElement.setAttribute('disabled', 'disabled');
+            if(rowElement){
+                rowElement.style = "display:none;";
+            }
+        }
+    }else{
+        finishedElement.innerText = "Marking file viewed!";
+        if(rowElement){
+            rowElement.style = "display:none;";
+        }
+    }
+}
+
 function markViewed(guid){
+    if(is_mcp){
+        _display_binge_redirect();
+        return;
+    }
+
     if(!viewed){
         viewed = true;
         jQuery.ajax({url: viewedUrl + guid + '/',
@@ -110,23 +158,7 @@ function markViewed(guid){
                 data: {viewed: 'true',
                        guid: guid},
                 success: function(json){
-                    var finishedElement = document.getElementById('viewedText');
-                    var toggleElement = document.getElementById('bingewatch-btn');
-                    var rowElement = document.getElementById('binge-watch-row');
-
-                    if(binge_mode && has_next_link && should_redirect){
-                        finishedElement.innerHTML = "Marking file viewed... Binge mode active! Preparing next video <button class='btn btn-info play-controls' name='cancel-btn' id='cancel-btn' onclick='cancelBingeWatch();'><i class='bi-x-octagon-fill'></i> Cancel</button>";
-                        if(toggleElement){
-                            toggleElement.onclick = function(){};
-                            toggleElement.setAttribute('disabled', 'disabled');
-                            rowElement.style = "display:none;";
-                        }
-                    }else{
-                        finishedElement.innerText = "Marking file viewed!";
-                        if(rowElement){
-                            rowElement.style = "display:none;";
-                        }
-                    }
+                    _display_binge_redirect()
                 }
             });
     }
